requirements.txt
dash==3.0.4
dash-bootstrap-components==2.0.2
pandas==2.2.3
plotly==6.0.1
gspread==6.2.1
google-auth==2.49.1
requests==2.32.3
scipy==1.17.1
pytz==2025.2
gunicorn==23.0.0
